+++ clevertap.js
@@ -6227,7 +6227,7 @@
     };
 
     const isExistingCampaign = campaignId => {
-      const testIframe = document.querySelector('iframe');
+      const testIframe = document.getElementById('wiz-iframe-intent') || document.getElementById('wiz-iframe');
 
       if (testIframe) {
         const iframeDocument = testIframe.contentDocument || testIframe.contentWindow.document;